refactor: Success에 NextTurn, Finished 이동
