+++ src/main/kotlin/omok/controller/OmokController.kt
@@ -5,9 +5,9 @@ import omok.domain.OmokBoard
 import omok.domain.OmokGame
 import omok.domain.Position
 import omok.domain.StoneState
-import omok.domain.turn.Finished
-import omok.domain.turn.NextTurn
 import omok.domain.turn.PutStoneResult
+import omok.domain.turn.PutStoneResult.Success.Finished
+import omok.domain.turn.PutStoneResult.Success.NextTurn
 import omok.domain.turn.TurnManager
 import omok.view.InputView
 import omok.view.OutputView
