--- src/main/kotlin/omok/domain/turn/NextTurn.kt
@@ -1,5 +1,0 @@
-package omok.domain.turn
-
-import omok.domain.StoneState
-
-class NextTurn(val turn: StoneState)
